docs: upgrade tanstack start, remove vinxi
