--- apps/docs/app/api.ts
@@ -1,6 +1,0 @@
-import {
-  createStartAPIHandler,
-  defaultAPIFileRouteHandler,
-} from '@tanstack/react-start/api';
-
-export default createStartAPIHandler(defaultAPIFileRouteHandler);
