--- apps/docs/app/client.tsx
@@ -1,8 +1,0 @@
-/// <reference types="vinxi/types/client" />
-import { StartClient } from '@tanstack/react-start';
-import { hydrateRoot } from 'react-dom/client';
-import { createRouter } from './router';
-
-const router = createRouter();
-
-hydrateRoot(document, <StartClient router={router} />);
