--- apps/docs/app/ssr.tsx
@@ -1,13 +1,0 @@
-/// <reference types="vinxi/types/server" />
-import { getRouterManifest } from '@tanstack/react-start/router-manifest';
-import {
-  createStartHandler,
-  defaultStreamHandler,
-} from '@tanstack/react-start/server';
-
-import { createRouter } from './router';
-
-export default createStartHandler({
-  createRouter,
-  getRouterManifest,
-})(defaultStreamHandler);
