+++ apps/docs/src/assets/obos_liggende_hus_svart_rgb.png
