+++ apps/docs/src/routes/__root.tsx
