+++ apps/docs/src/routes/api/health.ts
@@ -1,0 +1,7 @@
+import { createServerFileRoute } from '@tanstack/react-start/server';
+
+export const ServerRoute = createServerFileRoute('/api/health').methods({
+  GET: () => {
+    return new Response('Im alive!');
+  },
+});
