+++ apps/docs/src/ui/footer.tsx
