+++ apps/docs/src/ui/image-with-caption.tsx
