+++ apps/docs/tsconfig.json
@@ -9,7 +9,7 @@
     "baseUrl": "./",
     "noEmit": true,
     "paths": {
-      "@/*": ["app/*"]
+      "@/*": ["src/*"]
     }
   }
 }
